Add a new feature that the maxComment value is bound with the user such that each user can set their favorite maximum comments per page and the value will not change each time they login
Update the CommentResponse.java file to take in maxComment in type long instead of int.
Modify the AuthenticationServlet to adjust the welcome message shown on the screen.
Update the DataServlet.java file such that the maxComment value is stored in datastore and retrieved when the servlet handles a GET request.

+++ portfolio/src/main/java/com/google/sps/comment/CommentResponse.java
@@ -18,9 +18,9 @@ import java.util.List;
 
 public final class CommentResponse {
     private final List<Comment> commentList;
-    private final int maxComments;
+    private final long maxComments;
 
-    public CommentResponse(List<Comment> commentList, int maxComments) {
+    public CommentResponse(List<Comment> commentList, long maxComments) {
         this.commentList = commentList;
         this.maxComments = maxComments;
     }
